调试需求管理中的细节

+++ api/views_dir/permissions.py
@@ -14,7 +14,7 @@ import json
 # cerf  token验证 用户展示模块
 @csrf_exempt
 @account.is_token(models.userprofile)
-def company(request):
+def permissions(request):
     response = Response.ResponseObj()
     if request.method == "GET":
         forms_obj = SelectForm(request.GET)
@@ -28,10 +28,12 @@ def company(request):
                 'name': '__contains',
                 'create_date': '',
                 'oper_user__username': '__contains',
+                'pid_id': '__isnull'
             }
             q = conditionCom(request, field_dict)
             print('q -->', q)
-            objs = models.company.objects.filter(q).order_by(order)
+
+            objs = models.permissions.objects.select_related('pid').filter(q).order_by(order)
             count = objs.count()
 
             if length != 0:
@@ -50,9 +52,15 @@ def company(request):
                     oper_user_username = ''
                 # print('oper_user_username -->', oper_user_username)
                 #  将查询出来的数据 加入列表
+                pid_title = ''
+                if obj.pid:
+                    pid_title = obj.pid.title
                 ret_data.append({
                     'id': obj.id,
                     'name': obj.name,
+                    'title': obj.title,
+                    'pid_id': obj.pid_id,
+                    'pid_title': pid_title,
                     'create_date': obj.create_date.strftime('%Y-%m-%d %H:%M:%S'),
                     'oper_user__username': oper_user_username,
                 })
@@ -74,7 +82,7 @@ def company(request):
 #  csrf  token验证
 @csrf_exempt
 @account.is_token(models.userprofile)
-def company_oper(request, oper_type, o_id):
+def permissions_oper(request, oper_type, o_id):
     response = Response.ResponseObj()
     if request.method == "POST":
         if oper_type == "add":
@@ -82,6 +90,8 @@ def company_oper(request, oper_type, o_id):
                 'user_id': o_id,
                 'oper_user_id': request.GET.get('user_id'),
                 'name': request.POST.get('name'),
+                'title': request.POST.get('title'),
+                'pid_id': request.POST.get('pid_id'),
             }
             #  创建 form验证 实例（参数默认转成字典）
             forms_obj = AddForm(form_data)
@@ -90,7 +100,7 @@ def company_oper(request, oper_type, o_id):
                 # print(forms_obj.cleaned_data)
                 #  添加数据库
                 # print('forms_obj.cleaned_data-->',forms_obj.cleaned_data)
-                models.company.objects.create(**forms_obj.cleaned_data)
+                models.permissions.objects.create(**forms_obj.cleaned_data)
                 response.code = 200
                 response.msg = "添加成功"
             else:
@@ -102,11 +112,16 @@ def company_oper(request, oper_type, o_id):
 
         elif oper_type == "delete":
             # 删除 ID
-            objs = models.company.objects.filter(id=o_id)
+            objs = models.permissions.objects.filter(id=o_id)
             if objs:
-                objs.delete()
-                response.code = 200
-                response.msg = "删除成功"
+                obj = objs[0]
+                if models.permissions.objects.filter(pid_id=obj.id).count() > 0:
+                    response.code = 304
+                    response.msg = "含有子级数据,请先删除或转移子级数据"
+                else:
+                    objs.delete()
+                    response.code = 200
+                    response.msg = "删除成功"
             else:
                 response.code = 302
                 response.msg = '删除ID不存在'
@@ -115,6 +130,8 @@ def company_oper(request, oper_type, o_id):
             form_data = {
                 'o_id': o_id,
                 'name': request.POST.get('name'),
+                'title': request.POST.get('title'),
+                'pid_id': request.POST.get('pid_id'),
             }
 
             forms_obj = UpdateForm(form_data)
@@ -123,14 +140,18 @@ def company_oper(request, oper_type, o_id):
                 print(forms_obj.cleaned_data)
                 o_id = forms_obj.cleaned_data['o_id']
                 name = forms_obj.cleaned_data['name']
+                title = forms_obj.cleaned_data['title']
+                pid_id = forms_obj.cleaned_data['pid_id']
                 #  查询数据库  用户id
-                objs = models.company.objects.filter(
+                objs = models.permissions.objects.filter(
                     id=o_id
                 )
                 #  更新 数据
                 if objs:
                     objs.update(
-                        name=name
+                        name=name,
+                        title=title,
+                        pid_id=pid_id,
                     )
 
                     response.code = 200
